@@ -1,6 +1,6 @@
 {
   "name": "tr46",
-  "version": "5.0.0",
+  "version": "5.1.0",
   "engines": {
     "node": ">=18"
   },
@@ -12,7 +12,7 @@
   ],
   "scripts": {
     "test": "node --test",
-    "lint": "eslint .",
+    "lint": "eslint",
     "pretest": "node scripts/getLatestTests.js",
     "prepublish": "node scripts/generateMappingTable.js && node scripts/generateRegexes.js"
   },
@@ -34,10 +34,11 @@
     "punycode": "^2.3.1"
   },
   "devDependencies": {
-    "@domenic/eslint-config": "^3.0.0",
-    "@unicode/unicode-15.1.0": "^1.5.2",
-    "eslint": "^8.53.0",
+    "@domenic/eslint-config": "^4.0.1",
+    "@unicode/unicode-16.0.0": "^1.6.5",
+    "eslint": "^9.22.0",
+    "globals": "^16.0.0",
     "regenerate": "^1.4.2"
   },
-  "unicodeVersion": "15.1.0"
+  "unicodeVersion": "16.0.0"
 }
