@@ -1,6 +1,6 @@
 {
   "name": "whatwg-url",
-  "version": "14.1.1",
+  "version": "14.2.0",
   "description": "An implementation of the WHATWG URL Standard's URL API and parsing machinery",
   "main": "index.js",
   "files": [
@@ -12,16 +12,16 @@
   "license": "MIT",
   "repository": "jsdom/whatwg-url",
   "dependencies": {
-    "tr46": "^5.0.0",
+    "tr46": "^5.1.0",
     "webidl-conversions": "^7.0.0"
   },
   "devDependencies": {
     "@domenic/eslint-config": "^4.0.1",
     "benchmark": "^2.1.4",
-    "c8": "^10.1.2",
-    "esbuild": "^0.24.0",
-    "eslint": "^9.16.0",
-    "globals": "^15.13.0",
+    "c8": "^10.1.3",
+    "esbuild": "^0.25.1",
+    "eslint": "^9.22.0",
+    "globals": "^16.0.0",
     "webidl2js": "^18.0.0"
   },
   "engines": {
