@@ -1,16 +1,29 @@
 const mongoose = require('mongoose');
 
 const bookingSchema = new mongoose.Schema({
-  service: { type: mongoose.Schema.Types.ObjectId, ref: 'Service', required: true },
-  staff: { type: mongoose.Schema.Types.ObjectId, ref: 'Staff', required: true },
-  time: { type: String, required: true }, // e.g., "10:00 AM"
-  customerInfo: {
-    firstName: { type: String, required: true },
-    lastName: { type: String, required: true },
-    email: { type: String, required: true },
-    phoneNumber: { type: String, required: true }
+  serviceId: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'Service',
+    required: true,
   },
-  createdAt: { type: Date, default: Date.now }
+  staffId: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'Staff',
+    required: true,
+  },
+  timeSlot: {
+    type: Date,
+    required: true,
+  },
+  customerId: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'User',
+    required: true,
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
 });
 
 module.exports = mongoose.model('Booking', bookingSchema);
