@@ -1,9 +1,18 @@
 const mongoose = require('mongoose');
 
 const staffSchema = new mongoose.Schema({
-  name: { type: String, required: true },
-  role: { type: String, required: true },
-  bio: { type: String }
+  name: {
+    type: String,
+    required: true,
+  },
+  available: {
+    type: Boolean,
+    default: true,
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
 });
 
 module.exports = mongoose.model('Staff', staffSchema);
