@@ -1,8 +1,49 @@
 const mongoose = require('mongoose');
 
 const userSchema = new mongoose.Schema({
-  email: { type: String, required: true, unique: true },
-  password: { type: String, required: true }
+  name: {
+    type: String,
+    required: true,
+    trim: true,
+  },
+  email: {
+    type: String,
+    required: true,
+    unique: true,
+    trim: true,
+    lowercase: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+  accountType: {
+    type: String,
+    enum: ['customer', 'business'],
+    default: 'customer',
+    required: true,
+  },
+  businessName: {
+    type: String,
+    required: function() {
+      return this.accountType === 'business';
+    },
+  },
+  phoneNumber: {
+    type: String,
+  },
+  termsConsent: {
+    type: Boolean,
+    default: false,
+  },
+  marketingConsent: {
+    type: Boolean,
+    default: false,
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now,
+  },
 });
 
 module.exports = mongoose.model('User', userSchema);
