@@ -20,9 +20,6 @@ import { EmailVerificationComponent } from './pages/auth/email-verification/emai
 import { LoginComponent } from './pages/auth/login/login.component';
 import { PasswordResetComponent } from './pages/auth/password-reset/password-reset.component';
 import { SignupComponent } from './pages/auth/signup/signup.component';
-import { GoogleLoginComponent } from './pages/auth/google-login/google-login.component'; // Added import
-import { AppleLoginComponent } from './pages/auth/apple-login/apple-login.component'; // Added import
-import { FacebookLoginComponent } from './pages/auth/facebook-login/facebook-login.component'; // Added import
 import { AiConfigComponent } from './pages/admin/ai-config/ai-config.component';
 import { BrandingComponent } from './pages/admin/branding/branding.component';
 import { CalendarComponent } from './pages/admin/calendar/calendar.component';
@@ -50,38 +47,35 @@ export const routes: Routes = [
   { path: 'terms', component: TermsComponent },
   { path: 'cookie', component: CookieComponent },
 
-  // Booking Routes
-  { path: 'booking/service-selection', component: ServiceSelectionComponent },
-  { path: 'booking/staff-selection', component: StaffSelectionComponent },
-  { path: 'booking/time-slot-selection', component: TimeSlotSelectionComponent },
-  { path: 'booking/customer-info', component: CustomerInfoComponent },
-  { path: 'booking/confirmation', component: ConfirmationComponent },
-  { path: 'booking/payment', component: PaymentComponent },
-  { path: 'booking/success', component: SuccessComponent },
-  { path: 'booking/management', component: ManagementComponent },
+  // Booking Routes (Protected)
+  { path: 'booking/service-selection', component: ServiceSelectionComponent, canActivate: [authGuard] },
+  { path: 'booking/staff-selection', component: StaffSelectionComponent, canActivate: [authGuard] },
+  { path: 'booking/time-slot-selection', component: TimeSlotSelectionComponent, canActivate: [authGuard] },
+  { path: 'booking/customer-info', component: CustomerInfoComponent, canActivate: [authGuard] },
+  { path: 'booking/confirmation', component: ConfirmationComponent, canActivate: [authGuard] },
+  { path: 'booking/payment', component: PaymentComponent, canActivate: [authGuard] },
+  { path: 'booking/success', component: SuccessComponent, canActivate: [authGuard] },
+  { path: 'booking/management', component: ManagementComponent, canActivate: [authGuard] },
 
   // Auth Routes
   { path: 'verify-email', component: EmailVerificationComponent },
   { path: 'login', component: LoginComponent },
   { path: 'forgot-password', component: PasswordResetComponent },
-  { path: 'register', component: SignupComponent },
-  { path: 'auth/google', component: GoogleLoginComponent, canActivate: [authGuard] },
-  { path: 'auth/apple', component: AppleLoginComponent, canActivate: [authGuard] },
-  { path: 'auth/facebook', component: FacebookLoginComponent, canActivate: [authGuard] },
+  { path: 'signup', component: SignupComponent },
 
-  // Admin Routes
-  { path: 'admin/ai-config', component: AiConfigComponent },
-  { path: 'admin/branding', component: BrandingComponent },
-  { path: 'admin/calendar', component: CalendarComponent },
-  { path: 'admin/customers', component: CustomersComponent },
-  { path: 'admin/dashboard', component: DashboardComponent },
-  { path: 'admin/integrations', component: IntegrationsComponent },
-  { path: 'admin/notifications', component: NotificationsComponent },
-  { path: 'admin/reports', component: ReportsComponent },
-  { path: 'admin/services', component: ServicesComponent },
-  { path: 'admin/settings', component: SettingsComponent },
-  { path: 'admin/staff', component: StaffComponent },
-  { path: 'admin/system-health', component: SystemHealthComponent },
+  // Admin Routes (Protected)
+  { path: 'admin/ai-config', component: AiConfigComponent, canActivate: [authGuard] },
+  { path: 'admin/branding', component: BrandingComponent, canActivate: [authGuard] },
+  { path: 'admin/calendar', component: CalendarComponent, canActivate: [authGuard] },
+  { path: 'admin/customers', component: CustomersComponent, canActivate: [authGuard] },
+  { path: 'admin/dashboard', component: DashboardComponent, canActivate: [authGuard] },
+  { path: 'admin/integrations', component: IntegrationsComponent, canActivate: [authGuard] },
+  { path: 'admin/notifications', component: NotificationsComponent, canActivate: [authGuard] },
+  { path: 'admin/reports', component: ReportsComponent, canActivate: [authGuard] },
+  { path: 'admin/services', component: ServicesComponent, canActivate: [authGuard] },
+  { path: 'admin/settings', component: SettingsComponent, canActivate: [authGuard] },
+  { path: 'admin/staff', component: StaffComponent, canActivate: [authGuard] },
+  { path: 'admin/system-health', component: SystemHealthComponent, canActivate: [authGuard] },
 
   // Wildcard
   { path: '**', redirectTo: '/home', pathMatch: 'full' }
