@@ -1,57 +1,71 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { Auth, onAuthStateChanged, User } from '@angular/fire/auth';
 import { isPlatformBrowser } from '@angular/common';
 import { PLATFORM_ID, Inject } from '@angular/core';
-import { firstValueFrom, from } from 'rxjs';
 
-export const authGuard: CanActivateFn = async (route, state) => {
+export const authGuard: CanActivateFn = (route, state) => {
   const router = inject(Router);
   const platformId = inject(PLATFORM_ID);
-  let auth: Auth | undefined;
 
-  // Allow navigation to auth routes without checking authentication
-  const authRoutes = ['/auth/google', '/auth/apple', '/auth/facebook'];
-  if (authRoutes.includes(state.url)) {
-    console.log(`Allowing navigation to ${state.url} without auth check`);
+  return true; // TODO(Remove) Allow all routes for now till we implement the APIs
+
+  // Allow navigation to public and auth routes without checking authentication
+  const publicRoutes = [
+    '/',
+    '/home',
+    '/about',
+    '/contact',
+    '/features',
+    '/pricing',
+    '/faq',
+    '/privacy',
+    '/terms',
+    '/cookie',
+  ];
+  const authRoutes = ['/signup', '/login', '/forgot-password', '/verify-email'];
+  if (publicRoutes.includes(state.url) || authRoutes.includes(state.url)) {
+    console.log(`authGuard - Allowing navigation to ${state.url} without auth check`);
     return true;
   }
 
-  if (isPlatformBrowser(platformId)) {
-    try {
-      auth = inject(Auth);
-      console.log('Auth service in authGuard:', auth);
-    } catch (error) {
-      console.warn('Auth service is not available during guard execution:', error);
-    }
-  } else {
-    console.log('Running on server - skipping Auth injection in authGuard');
-  }
-
-  if (!auth) {
-    console.log('Redirecting to login due to missing Auth service');
+  if (!isPlatformBrowser(platformId)) {
+    console.log('authGuard - Running on server - redirecting to login');
     return router.createUrlTree(['/login']);
   }
 
-  // Check the user's authentication state for other routes
-  try {
-    const user = await firstValueFrom(
-      from(
-        new Promise<User | null>((resolve) => {
-          onAuthStateChanged(auth, (user) => resolve(user), (error) => resolve(null));
-        })
-      )
-    );
+  // Check for JWT token in localStorage
+  const isLoggedIn = localStorage.getItem('isLoggedIn') || 'false';
+  const currentUser = JSON.parse(localStorage.getItem('user') || '{}');
 
-    if (!user) {
-      console.log('User is not authenticated - redirecting to login');
+  if (!isLoggedIn || !currentUser || !currentUser.id) {
+    console.log('authGuard - No token or user found - redirecting to login');
+    return router.createUrlTree(['/login']);
+  }
+
+  // Role-based access control
+  const accountType = currentUser.accountType;
+
+  // Admin routes - only accessible to business users
+  if (state.url.startsWith('/admin')) {
+    if (accountType !== 'business') {
+      console.log('authGuard - User is not a business account - redirecting to home');
+      return router.createUrlTree(['/home']);
+    }
+    console.log('authGuard - Business user authenticated for admin route');
+    return true;
+  }
+
+  // Booking routes - accessible to authenticated users (both customer and business)
+  if (state.url.startsWith('/booking')) {
+    if (!accountType) {
+      console.log('authGuard - No account type found - redirecting to login');
       return router.createUrlTree(['/login']);
     }
-
-    console.log('User is authenticated:', user.uid);
+    console.log(`authGuard - ${accountType} user authenticated for booking route`);
     return true;
-  } catch (error) {
-    console.error('Error checking auth state:', error);
-    return router.createUrlTree(['/login']);
   }
+
+  // Default case - allow access if authenticated
+  console.log('authGuard - User is authenticated:', currentUser.id);
+  return true;
 };
