@@ -18,7 +18,7 @@ export class HeaderComponent implements OnInit {
     { label: 'FAQ', path: '/faq' },
     { label: 'Contact', path: '/contact' },
     { label: 'Sign In', path: '/login' },
-    { label: 'Sign Up', path: '/signup' }
+    { label: 'Sign Up', path: '/signup' } // Changed from '/register' to '/signup'
   ];
 
   isMobileNavActive = false;
