@@ -2,6 +2,7 @@ import { isPlatformBrowser } from '@angular/common';
 import { Inject, PLATFORM_ID, Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
+import { AdminService } from '../../../shared/services/admin.service';
 
 @Component({
   selector: 'app-settings',
@@ -10,26 +11,47 @@ import { CommonModule } from '@angular/common';
   templateUrl: './settings.component.html'
 })
 export class SettingsComponent implements OnInit {
-  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
-
-  settings: any = {};
+  constructor(
+    @Inject(PLATFORM_ID) private platformId: Object,
+    private adminService: AdminService
+  ) {}
 
   businessSettings = {
-    name: 'Nexora Salon',
-    hours: '9 AM - 5 PM',
-    location: '123 Main St'
+    name: '',
+    hours: '',
+    location: ''
   };
+  error: string | null = null;
 
   saveSettings(): void {
-    console.log('Settings Saved:', this.businessSettings);
+    this.adminService.updateBusinessSettings(this.businessSettings).subscribe({
+      next: () => {
+        alert('Settings Saved Successfully!');
+      },
+      error: (err) => {
+        this.error = err.error.message || 'Failed to save settings';
+        console.error('Failed to save settings:', err);
+      }
+    });
   }
 
   ngOnInit(): void {
     if (isPlatformBrowser(this.platformId)) {
       console.log('Running on browser');
+      const user = JSON.parse(localStorage.getItem('user') || '{}');
+      this.businessSettings.name = user.businessName || 'Nexora Salon';
+      this.adminService.getBusinessSettings().subscribe({
+        next: (settings) => {
+          this.businessSettings = { ...this.businessSettings, ...settings };
+        },
+        error: (err) => {
+          this.error = err.error.message || 'Failed to fetch settings';
+          console.error('Failed to fetch settings:', err);
+        }
+      });
     } else {
       console.log('Running on server');
-      this.settings = {};
+      this.businessSettings = { name: '', hours: '', location: '' };
     }
   }
 }
