@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ApiService } from './api.service';
 import { Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -9,11 +10,25 @@ export class AuthService {
   constructor(private api: ApiService) {}
 
   login(email: string, password: string): Observable<any> {
-    return this.api.post('auth/login', { email, password });
+    return this.api.post('auth/login', { email, password }).pipe(
+      tap(response => {
+        if (response.token && response.user) {
+          localStorage.setItem('token', response.token);
+          localStorage.setItem('user', JSON.stringify(response.user));
+        }
+      })
+    );
   }
 
-  signup(email: string, password: string, businessName?: string): Observable<any> {
-    return this.api.post('auth/signup', { email, password, businessName });
+  signup(fullName: string, email: string, password: string, accountType: string, marketingConsent: boolean): Observable<any> {
+    return this.api.post('auth/signup', { fullName, email, password, accountType, marketingConsent }).pipe(
+      tap(response => {
+        if (response.token && response.user) {
+          localStorage.setItem('token', response.token);
+          localStorage.setItem('user', JSON.stringify(response.user));
+        }
+      })
+    );
   }
 
   resetPassword(email: string): Observable<any> {
@@ -25,7 +40,16 @@ export class AuthService {
   }
 
   logout(): void {
-    // Placeholder for logout logic (e.g., clear local storage)
+    localStorage.removeItem('token');
+    localStorage.removeItem('user');
     console.log('User logged out');
+  }
+
+  isAuthenticated(): boolean {
+    return !!localStorage.getItem('token');
+  }
+
+  getUser(): any {
+    return JSON.parse(localStorage.getItem('user') || '{}');
   }
 }
