@@ -2,6 +2,7 @@ import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 import { ApiService } from './api.service';
 import { Observable, of } from 'rxjs';
 import { isPlatformBrowser } from '@angular/common';
+import { HttpHeaders } from '@angular/common/http';
 
 @Injectable({
   providedIn: 'root'
@@ -23,42 +24,54 @@ export class BookingService {
 
   getServices(): Observable<any[]> {
     if (isPlatformBrowser(this.platformId)) {
-      return this.api.get('bookings/services');
+      return this.api.get('services');
     }
     return of([]);
   }
 
   getStaff(): Observable<any[]> {
     if (isPlatformBrowser(this.platformId)) {
-      return this.api.get('bookings/staff');
+      return this.api.get('staff');
     }
     return of([]);
   }
 
-  getTimeSlots(date: string): Observable<any[]> {
+  getTimeSlots(staffId: string): Observable<any[]> {
     if (isPlatformBrowser(this.platformId)) {
-      return this.api.get(`bookings/time-slots?date=${date}`);
+      return this.api.get(`time-slots?staffId=${staffId}`);
     }
     return of([]);
   }
 
   createBooking(booking: any): Observable<any> {
     if (isPlatformBrowser(this.platformId)) {
-      return this.api.post('bookings/create', booking);
+      const token = localStorage.getItem('token');
+      const headers = new HttpHeaders({
+        Authorization: `Bearer ${token}`,
+      });
+      return this.api.post('bookings', booking, { headers });
     }
     return of({});
   }
 
   getBookings(): Observable<any[]> {
     if (isPlatformBrowser(this.platformId)) {
-      return this.api.get('bookings');
+      const token = localStorage.getItem('token');
+      const headers = new HttpHeaders({
+        Authorization: `Bearer ${token}`,
+      });
+      return this.api.get('bookings', { headers });
     }
     return of([]);
   }
 
-  cancelBooking(id: number): Observable<any> {
+  cancelBooking(bookingId: string): Observable<any> {
     if (isPlatformBrowser(this.platformId)) {
-      return this.api.post(`bookings/${id}/cancel`, {});
+      const token = localStorage.getItem('token');
+      const headers = new HttpHeaders({
+        Authorization: `Bearer ${token}`,
+      });
+      return this.api.delete(`bookings/${bookingId}`, { headers });
     }
     return of({});
   }
